+++ src/client/debug/debug.js
@@ -12,6 +12,7 @@ import Mousetrap from 'mousetrap';
 import { AssignShortcuts } from './assign-shortcuts';
 import { GameInfo } from './gameinfo';
 import { Controls } from './controls';
+import { PlayerInfo } from './playerinfo';
 import { DebugMove } from './debug-move';
 import { GameLog } from '../log/log';
 import { restore } from '../../core/action-creators';
@@ -31,11 +32,11 @@ export class Debug extends React.Component {
       ctx: PropTypes.any.isRequired,
       log: PropTypes.array.isRequired,
       isActive: PropTypes.bool,
+      isConnected: PropTypes.bool,
       _initial: PropTypes.any.isRequired,
     }),
     gameID: PropTypes.string.isRequired,
     playerID: PropTypes.string,
-    isConnected: PropTypes.bool,
     isMultiplayer: PropTypes.bool,
     moves: PropTypes.any,
     events: PropTypes.any,
@@ -47,11 +48,30 @@ export class Debug extends React.Component {
     reducer: PropTypes.func,
     overrideGameState: PropTypes.func,
     visualizeAI: PropTypes.func,
+    updateGameID: PropTypes.func,
+    updatePlayerID: PropTypes.func,
+    updateCredentials: PropTypes.func,
+    showGameInfo: PropTypes.bool,
+    dockControls: PropTypes.bool,
+  };
+
+  static defaultProps = {
+    showGameInfo: true,
+    dockControls: false,
   };
 
   constructor(props) {
     super(props);
     this.shortcuts = AssignShortcuts(props.moves, props.events, 'dlit');
+
+    this.state = {
+      showDebugUI: true,
+      showLog: false,
+      showGameInfo: props.showGameInfo,
+      dockControls: props.dockControls,
+      help: false,
+      AIMetadata: null,
+    };
   }
 
   componentDidMount() {
@@ -80,15 +100,6 @@ export class Debug extends React.Component {
     Mousetrap.unbind('d');
     Mousetrap.unbind('l');
   }
-
-  state = {
-    showDebugUI: true,
-    showLog: false,
-    showGameInfo: true,
-    dockControls: false,
-    help: false,
-    AIMetadata: null,
-  };
 
   saveState = () => {
     const json = JSON.stringify(this.props.gamestate);
@@ -155,19 +166,6 @@ export class Debug extends React.Component {
       );
     }
 
-    let players = [];
-    for (let i = 0; i < this.props.gamestate.ctx.numPlayers; i++) {
-      let className = 'player active';
-      if (i != this.props.gamestate.ctx.currentPlayer) {
-        className = 'player';
-      }
-      players.push(
-        <div className={className} key={i}>
-          {i}
-        </div>
-      );
-    }
-
     let className = 'debug-ui';
     if (this.state.dockControls) {
       className += ' docktop';
@@ -204,7 +202,7 @@ export class Debug extends React.Component {
                   gameID={this.props.gameID}
                   playerID={this.props.playerID}
                   isActive={this.props.gamestate.isActive}
-                  isConnected={this.props.isConnected}
+                  isConnected={this.props.gamestate.isConnected}
                   isMultiplayer={this.props.isMultiplayer}
                 />
               )}
@@ -221,7 +219,11 @@ export class Debug extends React.Component {
               />
 
               <h3>Players</h3>
-              <div className="player-box">{players}</div>
+              <PlayerInfo
+                ctx={this.props.gamestate.ctx}
+                playerID={this.props.playerID}
+                onClick={this.props.updatePlayerID}
+              />
 
               <h3>Moves</h3>